// src/app/shared/interfaces/employeeViewDetails.interface.ts
export class IEmployeeViewDetails{
    counter: number;
    empCode: string;
    employeeName: string;
    employeeId: string;
    strDateFrom: string;
    strDateTo: string;
    trainingName: string
    trainingHours: number;
    durationDays: number;
    vacationType: string;
    wfhFixedDay: string;
    trainingRemarks: string;
    vacationRemarks:string;
    
}