// src/app/shared/interfaces/employeeDataDetails.interface.ts
export class IEmployeeDataDetails{
    employeeCode: string;
    employeeTrainingCount: number;
    employeeVacationCount: number;
    employeeLeaveCount: number;
    employeeTravelCount: number;
    employeeWFHAdhocCount: number;
    allTraining = new Array();
    allLeave = new Array();
    allTravel = new Array();
    allWFHAdhoc = new Array();
    allWFHDays = new Array();
}